% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clusterdepth_array.R
\name{compute_clusterdepth_array}
\alias{compute_clusterdepth_array}
\title{Cluster Depth Tests}
\usage{
compute_clusterdepth_array(distribution, threshold, graph, alternative)
}
\arguments{
\item{distribution}{An 3d array representing the null distribution of multiple signal. The first dimension is the permutations, the second the samples, the third is the channels.}

\item{threshold}{A numeric specifying the height of the threshold to compute the clusters.}

\item{graph}{A igraph object representing the adjacency of the channels (used for the outputs only).}

\item{alternative}{A character string indicating the alternative hypothesis. Either \code{"greater"}, \code{"less"} or \code{"two.sided"}.}
}
\value{
A list containing an igraph and a data.frame, with the results for each sample, channel.
}
\description{
Compute the Cluster- multiple comparisons procedure.
}
\references{
Frossard, J., & Renaud, O. (2021). The Cluster Depth Tests: Toward Point-Wise Strong Control of the Family-Wise Error Rate in Massively Univariate Tests with Application to M/EEG. arXiv preprint arXiv:2105.07514.
}
\seealso{
Other MCP: 
\code{\link{compute_clustermass_array}()},
\code{\link{compute_stepdownmaxT_array}()},
\code{\link{compute_tfce_array}()},
\code{\link{compute_troendle_array}()}
}
\concept{MCP}
