% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_troendle_array.R
\name{compute_troendle_array}
\alias{compute_troendle_array}
\title{Troendle multiple comparison procedure}
\usage{
compute_troendle_array(distribution, alternative, graph)
}
\arguments{
\item{distribution}{An 3d array representing the null distribution of multiple signal. The first dimension is the permutations, the second the samples, the third is the channels.}

\item{alternative}{a character string indicating the alternative hypothesis. Either \code{"greater"}, \code{"less"} or \code{"two.sided"}.}

\item{graph}{A igraph object representing the adjacency of the channels.}
}
\value{
graph a list containing an igraph and a data.frame, with the results for each sample, channel.
}
\description{
Compute the Troendle multiple comparison procedure
}
\seealso{
Other MCP: 
\code{\link{compute_clustermass_array}()},
\code{\link{compute_tfce_array}()}
}
\concept{MCP}
