% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plot of coefficients by effect, for multiple channels through sample.}
\usage{
plot_coef(object, ...)
}
\arguments{
\item{object}{a formula or brainperm object}

\item{...}{other arguments including: \code{data} a dataframe of the design, \code{signal} a 3D array of the signal, \code{effect} a vector of the effect to plot, of \code{coding_sum}, a logical indicating if the effect should be computed using a contr.sum coding.}
}
\description{
Function to provide plot for exploratory analysis. Shows the effects (beta) through samples for several channels.
}
