% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_to_graph.R
\name{position_to_graph}
\alias{position_to_graph}
\title{Create igraph object from position of the channels}
\usage{
position_to_graph(data, delta = 4, name = "name", x = "x", y = "y", z = "z")
}
\arguments{
\item{data}{a data.frame containing the name and position of channels.}

\item{delta}{an double defining the maximal distance for adjacency of two channels.}

\item{name}{the colomn in the data containing the name of the channels.}

\item{x}{the colomn in the data containing the X position of the channels.}

\item{y}{the colomn in the data containing the Y position of the channels.}

\item{z}{the colomn in the data containing the Z position of the channels.}
}
\description{
Create an igraph object defining adjacency of the channels. Adjacency is defined when 2 channels have a (EUclidean) distance inferior to delta.
}
