% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.brainperm.R
\name{print.brainperm}
\alias{print.brainperm}
\title{Print method for \code{brainperm} object}
\usage{
\method{print}{brainperm}(x, ...)
}
\arguments{
\item{x}{a \code{brainperm} object.}

\item{...}{others arguments pass to code{summary.brainperm}.}
}
\description{
Display the results of the \code{brainperm} object.
}
\seealso{
summary.brainperm

Other summary functions: 
\code{\link{summary.brainperm}()}
}
\concept{summary functions}
