% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.brainperm.R
\name{summary.brainperm}
\alias{summary.brainperm}
\title{Print method for \code{brainperm} object}
\usage{
\method{summary}{brainperm}(
  object,
  multcomp = NULL,
  table_type = "cluster",
  alternative = "two.sided",
  ...
)
}
\arguments{
\item{object}{a \code{brainperm} object.}

\item{multcomp}{a character string indicating the multiple comparison procedure to display. Use if several multiple comparison procedures is computed in the \code{brainperm} object.}

\item{table_type}{a character string indicating the type of table. The default value, \code{table_type = "cluster"}, displays results by clusters. If \code{table_type = "full"}, the table shows the results by individual tests.}

\item{alternative}{a character string indicating the alternative hypothesis for t-test. Default value \code{"two.sided"}. \code{"less"} and \code{"greater"} is also available.}

\item{...}{other arguments pass to print.}
}
\description{
Display the results of the \code{brainperm} object.
}
