% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clusterdepth_array.R
\name{compute_clusterdepth_array}
\alias{compute_clusterdepth_array}
\title{Cluster-depth Tests}
\usage{
compute_clusterdepth_array(distribution, threshold, graph, alternative, border)
}
\arguments{
\item{distribution}{An 3d array representing the null distribution of multiple signal. The first dimension is the permutations, the second the samples, the third is the channels.}

\item{threshold}{A numeric specifying the height of the threshold to compute the clusters.}

\item{graph}{A igraph object representing the adjacency of the channels.}

\item{alternative}{a character string indicating the alternative hypothesis. Either \code{"greater"}, \code{"less"} or \code{"two.sided"}.}

\item{border}{A character string indicating the method to handles the border. Either \code{"reverse"}, reversing the computation of the cluster-depth for clusters at the border of the time frame, or \code{"ignore"}.}
}
\value{
graph a list containing an igraph and a data.frame, with the results for each sample, channel.
}
\description{
Compute the Cluster- multiple comparison procedure
}
\seealso{
Other MCP: 
\code{\link{compute_clustermass_array}()},
\code{\link{compute_stepdownmaxT_array}()},
\code{\link{compute_tfce_array}()},
\code{\link{compute_troendle_array}()}
}
\concept{MCP}
