% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clustermass_array.R
\name{compute_clustermass_array}
\alias{compute_clustermass_array}
\title{Cluster-mass test}
\usage{
compute_clustermass_array(
  distribution,
  threshold,
  aggr_FUN,
  graph,
  alternative
)
}
\arguments{
\item{distribution}{An 3d array representing the null distribution of multiple signal. The first dimension is the permutations, the second the samples, the third is the channels.}

\item{threshold}{The threshold used to compute the clusters.}

\item{aggr_FUN}{The function that aggregate the cluster into a scalar (cluster mass).}

\item{graph}{A igraph object representing the adjacency of the channels.}

\item{alternative}{a character string indicating the alternative hypothesis. Either \code{"greater"}, \code{"less"} or \code{"two.sided"}.}
}
\value{
graph A list containing a igraph object, a data frame containing the channels, time, statistic, cluster-mass and p-values,
}
\description{
Compute the cluster-mass test with adjacency define by a graph.
}
\seealso{
Other MCP: 
\code{\link{compute_tfce_array}()},
\code{\link{compute_troendle_array}()}
}
\concept{MCP}
