% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.brainperm.R
\name{plot.brainperm}
\alias{plot.brainperm}
\title{Plot a graph of the channels for 1 sample/time-point}
\usage{
\method{plot}{brainperm}(x, effect = 1, samples, ...)
}
\arguments{
\item{x}{brainperm object with a clustermass test.}

\item{effect}{an integer specifying which effect to plot.}

\item{samples}{a vector of integers specifying the samples to plots.}

\item{...}{other arguments including: \code{effect}: an integer specifying the effect to plot, \code{alpha}: argument to pass in par() and argument to pass in plot().}
}
\value{
a plot of a graph of the channels.
}
\description{
Plot the graphs of the channels for several samples. Red indicates significant clusters, grey non-significant and white represent below the threshold.
}
\seealso{
Other plotting functions: 
\code{\link{image.brainperm}()}
}
\concept{plotting functions}
