% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_to_graph.R
\name{position_to_graph}
\alias{position_to_graph}
\title{Create a graph of adjacency based on the position of the channel}
\usage{
position_to_graph(data, delta = 4, name = "name", x = "x", y = "y", z = "z")
}
\arguments{
\item{data}{a data.frame containing the name and position of channels.}

\item{delta}{a numeric defining the maximal distance for adjacency of two channels.}

\item{name}{the column in the data containing the name of the channels.}

\item{x}{the column in the data containing the X position of the channels.}

\item{y}{the column in the data containing the Y position of the channels.}

\item{z}{the column in the data containing the Z position of the channels.}
}
\value{
an \code{igraph} object.
}
\description{
Create an igraph object defining adjacency of the channels. Adjacency is defined when 2 channels have a (Euclidean) distance inferior to delta.
}
\concept{helper functions}
