% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.brainperm.R
\name{summary.brainperm}
\alias{summary.brainperm}
\title{Summary method for \code{brainperm} object}
\usage{
\method{summary}{brainperm}(
  object,
  multcomp = NULL,
  table_type = "cluster",
  alternative = "two.sided",
  ...
)
}
\arguments{
\item{object}{a \code{brainperm} object.}

\item{multcomp}{a character string indicating the multiple comparison procedure to display. Only relevant if several multiple comparison procedures are computed in the \code{brainperm} object.}

\item{table_type}{a character string indicating the type of table. The default value, \code{table_type = "cluster"}, displays results by clusters. If \code{table_type = "full"}, the table shows the results by individual tests.}

\item{alternative}{a character string indicating the alternative hypothesis for t-test. The default value if \code{"two.sided"}. \code{"less"} and \code{"greater"} are also available.}

\item{...}{other arguments pass to print.}
}
\value{
a list of tables containing the results of the test for each effect.
}
\description{
List per effect of the results of the \code{brainperm} object.
}
\seealso{
Other summary functions: 
\code{\link{print.brainperm}()}
}
\concept{summary functions}
